refactor: Move to SWC for backend
