+++ package.json
@@ -72,8 +72,8 @@
   },
   "devDependencies": {
     "@kayahr/jest-electron-runner": "^29.3.1",
-    "@swc/cli": "0.1.61",
-    "@swc/core": "1.3.35",
+    "@swc/cli": "0.3.12",
+    "@swc/core": "1.4.11",
     "@types/follow-redirects": "1.8.0",
     "@types/fs-extra": "8.1.0",
     "@types/jest": "27.5.1",
@@ -115,7 +115,6 @@
     "ts-transform-paths": "^2.0.3",
     "tsconfig-paths-webpack-plugin": "3.2.0",
     "tslint": "5.18.0",
-    "tswc": "^1.1.1",
     "ttypescript": "^1.5.15",
     "typescript": "4.6.2",
     "webpack": "5.90.3",
